exhibition-ingestion-module
