+++ .gitignore
@@ -1,0 +1,42 @@
+.venv/
+venv/
+ENV/
+
+__pycache__/
+*.py[cod]
+*$py.class
+*.so
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+
+.idea/
+.vscode/
+*.swp
+*.swo
+.DS_Store
+
+.env
+.env.local
+.env.*.local
+
+*.log
+
+*.sqlite3
+*.db
+
+.cache/
+.pytest_cache/
